Extract Snapshot layout to common components
